Revert "Merge pull request #84313 from fedirjh/sync-card-feed-on-demand"
This reverts commit 926a0753f556524babad3a919d908317d62cb87a, reversing
changes made to 5814d5bce3ee4b9a9e62afd9ea84ac6bbf8c749b.
