+++ src/languages/es.ts
@@ -5262,10 +5262,6 @@ ${amount} para ${merchant} - ${date}`,
                 removeCardFeed: 'Quitar la alimentación de tarjetas',
                 removeCardFeedTitle: (feedName) => `Eliminar el feed de ${feedName}`,
                 removeCardFeedDescription: '¿Estás seguro de que deseas eliminar esta fuente de tarjetas? Esto anulará la asignación de todas las tarjetas.',
-                assignNewCards: 'Asignar nuevas tarjetas',
-                assignNewCardsDescription: 'Obtén las últimas tarjetas de tu banco para asignar',
-                refreshConnectionSuccess: 'Conexión actualizada',
-                refreshConnectionSuccessDescription: 'La conexión con tu banco ha sido reautenticada exitosamente. Ahora puedes asignar nuevas tarjetas.',
                 error: {
                     feedNameRequired: 'Se requiere el nombre de la fuente de la tarjeta',
                     statementCloseDateRequired: 'Por favor, selecciona una fecha de cierre del estado de cuenta.',
