+++ src/languages/zh-hans.ts
@@ -5236,10 +5236,6 @@ _如需更详细的说明，请[访问我们的帮助网站](${CONST.NETSUITE_IM
                 removeCardFeed: '移除卡片流水',
                 removeCardFeedTitle: (feedName: string) => `移除 ${feedName} 数据源`,
                 removeCardFeedDescription: '确定要移除此卡片数据源吗？这将取消分配所有卡片。',
-                assignNewCards: '分配新卡片',
-                assignNewCardsDescription: '从银行获取最新的卡片进行分配',
-                refreshConnectionSuccess: '连接已刷新',
-                refreshConnectionSuccessDescription: '您的银行连接已成功重新验证。您现在可以分配新卡片。',
                 error: {
                     feedNameRequired: '必须填写卡片流水名称',
                     statementCloseDateRequired: '请选择账单结算日期。',
