+++ src/pages/workspace/companyCards/BankConnection/index.native.tsx
@@ -1,27 +1,35 @@
-import React, {useRef, useState} from 'react';
+import React, {useEffect, useMemo, useRef, useState} from 'react';
 import type {WebViewNavigation} from 'react-native-webview';
 import {WebView} from 'react-native-webview';
 import ActivityIndicator from '@components/ActivityIndicator';
 import FullPageOfflineBlockingView from '@components/BlockingViews/FullPageOfflineBlockingView';
-import ConfirmationPage from '@components/ConfirmationPage';
 import FullScreenLoadingIndicator from '@components/FullscreenLoadingIndicator';
 import HeaderWithBackButton from '@components/HeaderWithBackButton';
 import ScreenWrapper from '@components/ScreenWrapper';
+import useCardFeeds from '@hooks/useCardFeeds';
+import useImportPlaidAccounts from '@hooks/useImportPlaidAccounts';
+import useIsBlockedToAddFeed from '@hooks/useIsBlockedToAddFeed';
 import useLocalize from '@hooks/useLocalize';
 import useOnyx from '@hooks/useOnyx';
+import usePrevious from '@hooks/usePrevious';
 import useThemeStyles from '@hooks/useThemeStyles';
+import useUpdateFeedBrokenConnection from '@hooks/useUpdateFeedBrokenConnection';
+import {updateSelectedFeed} from '@libs/actions/Card';
+import {setAssignCardStepAndData} from '@libs/actions/CompanyCards';
+import {checkIfNewFeedConnected, getBankName, getCompanyCardFeed, isSelectedFeedExpired} from '@libs/CardUtils';
 import getUAForWebView from '@libs/getUAForWebView';
 import Navigation from '@libs/Navigation/Navigation';
 import type {SkeletonSpanReasonAttributes} from '@libs/telemetry/useSkeletonSpan';
 import type {PlatformStackRouteProp} from '@navigation/PlatformStackNavigation/types';
 import type {SettingsNavigatorParamList} from '@navigation/types';
 import WorkspaceCompanyCardsErrorConfirmation from '@pages/workspace/companyCards/WorkspaceCompanyCardsErrorConfirmation';
+import {setAddNewCompanyCardStepAndData} from '@userActions/CompanyCards';
+import {getCompanyCardBankConnection} from '@userActions/getCompanyCardBankConnection';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
 import type SCREENS from '@src/SCREENS';
 import type {CompanyCardFeedWithDomainID} from '@src/types/onyx';
-import useBankConnection from './useBankConnection';
 
 type BankConnectionProps = {
     /** ID of the policy */
@@ -32,43 +40,37 @@ type BankConnectionProps = {
 
     /** Route params for add new card flow */
     route?: PlatformStackRouteProp<SettingsNavigatorParamList, typeof SCREENS.WORKSPACE.COMPANY_CARDS_BANK_CONNECTION>;
-
-    /** Whether this is a refresh card list flow */
-    isRefreshConnectionFlow?: boolean;
-
-    /** Called when the assign flow succeeds */
-    onSuccess?: () => void;
-
-    /** Called when the assign flow fails due to broken connection */
-    onFailure?: () => void;
-
-    /** Called when the back button is pressed */
-    onBackButtonPress?: () => void;
 };
 
-function BankConnection({policyID: policyIDFromProps, feed, route, isRefreshConnectionFlow, onSuccess, onFailure, onBackButtonPress}: BankConnectionProps) {
+function BankConnection({policyID: policyIDFromProps, feed, route}: BankConnectionProps) {
     const styles = useThemeStyles();
     const {translate} = useLocalize();
     const webViewRef = useRef<WebView>(null);
     const [session] = useOnyx(ONYXKEYS.SESSION);
+    const [assignCard] = useOnyx(ONYXKEYS.ASSIGN_CARD);
     const authToken = session?.authToken ?? null;
+    const [addNewCard] = useOnyx(ONYXKEYS.ADD_NEW_COMPANY_CARD);
+    const selectedBank = addNewCard?.data?.selectedBank;
     const {feed: bankNameFromRoute, backTo, policyID: policyIDFromRoute} = route?.params ?? {};
     const policyID = policyIDFromProps ?? policyIDFromRoute;
+    const bankName = feed ? getBankName(getCompanyCardFeed(feed)) : (bankNameFromRoute ?? addNewCard?.data?.plaidConnectedFeed ?? selectedBank);
+    const plaidToken = addNewCard?.data?.publicToken ?? assignCard?.cardToAssign?.plaidAccessToken;
+    const isPlaid = !!plaidToken;
+    const url = getCompanyCardBankConnection(policyID, bankName);
+    const [cardFeeds] = useCardFeeds(policyID);
     const [isConnectionCompleted, setConnectionCompleted] = useState(false);
-
-    const {handleBackButtonPress, url, isPlaid, isNewFeedHasError, newFeed, isAllFeedsResultLoading, isBlockedToAddNewFeeds, isRefreshComplete} = useBankConnection({
-        policyID,
-        feed,
-        bankNameFromRoute,
-        onSuccess,
-        onFailure,
-        onBackButtonPress,
-        isRefreshConnectionFlow,
-        shouldOpenWindow: false,
-    });
-
+    const prevFeedsData = usePrevious(cardFeeds);
+    const isFeedExpired = feed ? isSelectedFeedExpired(cardFeeds?.[feed]) : false;
+    const {isNewFeedConnected, newFeed} = useMemo(
+        () => checkIfNewFeedConnected(prevFeedsData ?? {}, cardFeeds ?? {}, addNewCard?.data?.plaidConnectedFeed),
+        [addNewCard?.data?.plaidConnectedFeed, cardFeeds, prevFeedsData],
+    );
     const headerTitleAddCards = !backTo ? translate('workspace.companyCards.addCards') : undefined;
-    const headerTitle = feed ? translate(isRefreshConnectionFlow ? 'workspace.moreFeatures.companyCards.assignNewCards' : 'workspace.companyCards.assignCard') : headerTitleAddCards;
+    const headerTitle = feed ? translate('workspace.companyCards.assignCard') : headerTitleAddCards;
+    const onImportPlaidAccounts = useImportPlaidAccounts(policyID);
+    const {updateBrokenConnection, isFeedConnectionBroken} = useUpdateFeedBrokenConnection({policyID, feed});
+    const isNewFeedHasError = !!(newFeed && cardFeeds?.[newFeed]?.errors);
+    const {isBlockedToAddNewFeeds, isAllFeedsResultLoading} = useIsBlockedToAddFeed(policyID);
 
     const fullscreenReasonAttributes: SkeletonSpanReasonAttributes = {
         context: 'BankConnection',
@@ -82,6 +84,83 @@ function BankConnection({policyID: policyIDFromProps, feed, route, isRefreshConn
     };
     const renderLoading = () => <FullScreenLoadingIndicator reasonAttributes={fullscreenReasonAttributes} />;
 
+    useEffect(() => {
+        if (!policyID || !isBlockedToAddNewFeeds || feed) {
+            return;
+        }
+        Navigation.navigate(ROUTES.WORKSPACE_UPGRADE.getRoute(policyID, CONST.UPGRADE_FEATURE_INTRO_MAPPING.companyCards.alias, ROUTES.WORKSPACE_COMPANY_CARDS.getRoute(policyID)), {
+            forceReplace: true,
+        });
+    }, [isBlockedToAddNewFeeds, policyID, feed]);
+
+    const handleBackButtonPress = () => {
+        // Handle assign card flow
+        if (feed) {
+            Navigation.goBack();
+            return;
+        }
+
+        // Handle add new card flow
+        if (backTo) {
+            Navigation.goBack(backTo);
+            return;
+        }
+        setAddNewCompanyCardStepAndData({step: CONST.COMPANY_CARDS.STEP.SELECT_BANK});
+    };
+
+    useEffect(() => {
+        if ((!url && !isPlaid) || isNewFeedHasError) {
+            return;
+        }
+
+        // Handle assign card flow
+        if (feed && !isFeedExpired) {
+            if (isFeedConnectionBroken) {
+                updateBrokenConnection();
+                Navigation.goBack(ROUTES.WORKSPACE_COMPANY_CARDS.getRoute(policyID));
+                return;
+            }
+            setAssignCardStepAndData({
+                currentStep: assignCard?.cardToAssign?.dateOption ? CONST.COMPANY_CARD.STEP.CONFIRMATION : CONST.COMPANY_CARD.STEP.ASSIGNEE,
+                isEditing: false,
+            });
+            return;
+        }
+
+        // Handle add new card flow
+        if (isNewFeedConnected) {
+            if (newFeed) {
+                updateSelectedFeed(newFeed, policyID);
+            }
+
+            // Direct feeds (except those added via Plaid) are created with default statement period end date.
+            // Redirect the user to set a custom date.
+            if (policyID && !isPlaid) {
+                setAddNewCompanyCardStepAndData({
+                    step: CONST.COMPANY_CARDS.STEP.SELECT_DIRECT_STATEMENT_CLOSE_DATE,
+                });
+            } else {
+                Navigation.goBack(ROUTES.WORKSPACE_COMPANY_CARDS.getRoute(policyID));
+            }
+        }
+        if (isPlaid) {
+            onImportPlaidAccounts();
+        }
+    }, [
+        isNewFeedConnected,
+        newFeed,
+        policyID,
+        url,
+        feed,
+        isFeedExpired,
+        assignCard?.cardToAssign?.dateOption,
+        isPlaid,
+        onImportPlaidAccounts,
+        isFeedConnectionBroken,
+        updateBrokenConnection,
+        isNewFeedHasError,
+    ]);
+
     const checkIfConnectionCompleted = (navState: WebViewNavigation) => {
         if (!navState.url.includes(ROUTES.BANK_CONNECTION_COMPLETE)) {
             return;
@@ -89,56 +168,9 @@ function BankConnection({policyID: policyIDFromProps, feed, route, isRefreshConn
         setConnectionCompleted(true);
     };
 
-    const getContent = () => {
-        if (isRefreshComplete) {
-            return (
-                <ConfirmationPage
-                    heading={translate('workspace.moreFeatures.companyCards.refreshConnectionSuccess')}
-                    description={translate('workspace.moreFeatures.companyCards.refreshConnectionSuccessDescription')}
-                    shouldShowButton
-                    buttonText={translate('common.buttonConfirm')}
-                    onButtonPress={() => Navigation.dismissModal()}
-                />
-            );
-        }
-        if (isNewFeedHasError) {
-            return (
-                <WorkspaceCompanyCardsErrorConfirmation
-                    policyID={policyID}
-                    newFeed={newFeed}
-                />
-            );
-        }
-        if (!!url && !isConnectionCompleted && !isPlaid && !isAllFeedsResultLoading && (!isBlockedToAddNewFeeds || !!feed)) {
-            return (
-                <WebView
-                    ref={webViewRef}
-                    source={{
-                        uri: url,
-                        headers: {
-                            Cookie: `authToken=${authToken}`,
-                        },
-                    }}
-                    userAgent={getUAForWebView()}
-                    incognito
-                    onNavigationStateChange={checkIfConnectionCompleted}
-                    startInLoadingState
-                    renderLoading={renderLoading}
-                />
-            );
-        }
-        return (
-            <ActivityIndicator
-                size={CONST.ACTIVITY_INDICATOR_SIZE.LARGE}
-                style={styles.flex1}
-                reasonAttributes={activityReasonAttributes}
-            />
-        );
-    };
-
     return (
         <ScreenWrapper
-            testID={isRefreshComplete ? 'RefreshCardFeedConnectionSuccess' : 'BankConnection'}
+            testID="BankConnection"
             shouldShowOfflineIndicator={false}
             shouldEnablePickerAvoiding={false}
             shouldEnableMaxHeight
@@ -147,7 +179,37 @@ function BankConnection({policyID: policyIDFromProps, feed, route, isRefreshConn
                 title={headerTitle}
                 onBackButtonPress={handleBackButtonPress}
             />
-            <FullPageOfflineBlockingView addBottomSafeAreaPadding>{getContent()}</FullPageOfflineBlockingView>
+            <FullPageOfflineBlockingView addBottomSafeAreaPadding>
+                {!!url && !isConnectionCompleted && !isPlaid && !isNewFeedHasError && !isAllFeedsResultLoading && (!isBlockedToAddNewFeeds || !!feed) && (
+                    <WebView
+                        ref={webViewRef}
+                        source={{
+                            uri: url,
+                            headers: {
+                                Cookie: `authToken=${authToken}`,
+                            },
+                        }}
+                        userAgent={getUAForWebView()}
+                        incognito
+                        onNavigationStateChange={checkIfConnectionCompleted}
+                        startInLoadingState
+                        renderLoading={renderLoading}
+                    />
+                )}
+                {(isAllFeedsResultLoading || (isBlockedToAddNewFeeds && !feed) || isConnectionCompleted || isPlaid) && !isNewFeedHasError && (
+                    <ActivityIndicator
+                        size={CONST.ACTIVITY_INDICATOR_SIZE.LARGE}
+                        style={styles.flex1}
+                        reasonAttributes={activityReasonAttributes}
+                    />
+                )}
+                {isNewFeedHasError && (
+                    <WorkspaceCompanyCardsErrorConfirmation
+                        policyID={policyID}
+                        newFeed={newFeed}
+                    />
+                )}
+            </FullPageOfflineBlockingView>
         </ScreenWrapper>
     );
 }
