+++ src/pages/workspace/companyCards/BankConnection/index.tsx
@@ -1,24 +1,39 @@
-import React from 'react';
+import React, {useCallback, useEffect, useMemo, useState} from 'react';
 import ActivityIndicator from '@components/ActivityIndicator';
 import BlockingView from '@components/BlockingViews/BlockingView';
 import FullPageOfflineBlockingView from '@components/BlockingViews/FullPageOfflineBlockingView';
-import ConfirmationPage from '@components/ConfirmationPage';
 import HeaderWithBackButton from '@components/HeaderWithBackButton';
 import ScreenWrapper from '@components/ScreenWrapper';
 import Text from '@components/Text';
 import TextLink from '@components/TextLink';
+import useCardFeeds from '@hooks/useCardFeeds';
+import useImportPlaidAccounts from '@hooks/useImportPlaidAccounts';
+import useIsBlockedToAddFeed from '@hooks/useIsBlockedToAddFeed';
 import {useMemoizedLazyIllustrations} from '@hooks/useLazyAsset';
 import useLocalize from '@hooks/useLocalize';
+import useNetwork from '@hooks/useNetwork';
+import useOnyx from '@hooks/useOnyx';
+import usePrevious from '@hooks/usePrevious';
 import useThemeStyles from '@hooks/useThemeStyles';
+import useUpdateFeedBrokenConnection from '@hooks/useUpdateFeedBrokenConnection';
+import {setAssignCardStepAndData} from '@libs/actions/CompanyCards';
+import {checkIfNewFeedConnected, getBankName, getCompanyCardFeed, isSelectedFeedExpired} from '@libs/CardUtils';
 import Navigation from '@libs/Navigation/Navigation';
 import type {SkeletonSpanReasonAttributes} from '@libs/telemetry/useSkeletonSpan';
 import type {PlatformStackRouteProp} from '@navigation/PlatformStackNavigation/types';
 import type {SettingsNavigatorParamList} from '@navigation/types';
 import WorkspaceCompanyCardsErrorConfirmation from '@pages/workspace/companyCards/WorkspaceCompanyCardsErrorConfirmation';
+import {updateSelectedFeed} from '@userActions/Card';
+import {setAddNewCompanyCardStepAndData} from '@userActions/CompanyCards';
+import {getCompanyCardBankConnection} from '@userActions/getCompanyCardBankConnection';
 import CONST from '@src/CONST';
+import ONYXKEYS from '@src/ONYXKEYS';
+import ROUTES from '@src/ROUTES';
 import type SCREENS from '@src/SCREENS';
 import type {CompanyCardFeedWithDomainID} from '@src/types/onyx';
-import useBankConnection from './useBankConnection';
+import openBankConnection from './openBankConnection';
+
+let customWindow: Window | null = null;
 
 type BankConnectionProps = {
     /** ID of the policy */
@@ -29,70 +44,156 @@ type BankConnectionProps = {
 
     /** Route params for add new card flow */
     route?: PlatformStackRouteProp<SettingsNavigatorParamList, typeof SCREENS.WORKSPACE.COMPANY_CARDS_BANK_CONNECTION>;
-
-    /** Whether this is a refresh card list flow */
-    isRefreshConnectionFlow?: boolean;
-
-    /** Called when the assign flow succeeds */
-    onSuccess?: () => void;
-
-    /** Called when the assign flow fails due to broken connection */
-    onFailure?: () => void;
-
-    /** Called when the back button is pressed */
-    onBackButtonPress?: () => void;
 };
 
-function BankConnection({policyID: policyIDFromProps, feed, route, isRefreshConnectionFlow, onSuccess, onFailure, onBackButtonPress}: BankConnectionProps) {
+function BankConnection({policyID: policyIDFromProps, feed, route}: BankConnectionProps) {
     const styles = useThemeStyles();
     const {translate} = useLocalize();
-    const illustrations = useMemoizedLazyIllustrations(['PendingBank']);
+    const [addNewCard] = useOnyx(ONYXKEYS.ADD_NEW_COMPANY_CARD);
+    const [assignCard] = useOnyx(ONYXKEYS.ASSIGN_CARD);
     const {feed: bankNameFromRoute, backTo, policyID: policyIDFromRoute} = route?.params ?? {};
     const policyID = policyIDFromProps ?? policyIDFromRoute;
+    const [cardFeeds] = useCardFeeds(policyID);
+    const prevFeedsData = usePrevious(cardFeeds);
+    const illustrations = useMemoizedLazyIllustrations(['PendingBank']);
+    const [shouldBlockWindowOpen, setShouldBlockWindowOpen] = useState(false);
+    const selectedBank = addNewCard?.data?.selectedBank;
+    const bankName = feed ? getBankName(getCompanyCardFeed(feed)) : (bankNameFromRoute ?? addNewCard?.data?.plaidConnectedFeed ?? selectedBank);
+    const {isNewFeedConnected, newFeed} = useMemo(
+        () => checkIfNewFeedConnected(prevFeedsData ?? {}, cardFeeds ?? {}, addNewCard?.data?.plaidConnectedFeed),
+        [addNewCard?.data?.plaidConnectedFeed, cardFeeds, prevFeedsData],
+    );
+    const {isOffline} = useNetwork();
+    const plaidToken = addNewCard?.data?.publicToken ?? assignCard?.cardToAssign?.plaidAccessToken;
+    const {updateBrokenConnection, isFeedConnectionBroken} = useUpdateFeedBrokenConnection({policyID, feed});
+    const isPlaid = !!plaidToken;
 
-    const {
-        onOpenBankConnectionFlow,
-        handleBackButtonPress,
-        bankName,
-        bankDisplayName,
-        isPlaid,
-        isNewFeedHasError,
-        newFeed,
-        isAllFeedsResultLoading,
-        isBlockedToAddNewFeeds,
-        isRefreshComplete,
-    } = useBankConnection({
-        policyID,
-        feed,
-        bankNameFromRoute,
-        onSuccess,
-        onFailure,
-        onBackButtonPress,
-        isRefreshConnectionFlow,
-    });
+    const url = getCompanyCardBankConnection(policyID, bankName);
+    const isFeedExpired = feed ? isSelectedFeedExpired(cardFeeds?.[feed]) : false;
+    const headerTitleAddCards = !backTo ? translate('workspace.companyCards.addCards') : undefined;
+    const headerTitle = feed ? translate('workspace.companyCards.assignCard') : headerTitleAddCards;
+    const isNewFeedHasError = !!(newFeed && cardFeeds?.[newFeed]?.errors);
+    const onImportPlaidAccounts = useImportPlaidAccounts(policyID);
+    const {isBlockedToAddNewFeeds, isAllFeedsResultLoading} = useIsBlockedToAddFeed(policyID);
 
-    const headerTitleAddCards = !backTo ? translate(isRefreshConnectionFlow ? 'workspace.moreFeatures.companyCards.assignNewCards' : 'workspace.companyCards.addCards') : undefined;
-    const headerTitle = feed ? translate(isRefreshConnectionFlow ? 'workspace.moreFeatures.companyCards.assignNewCards' : 'workspace.companyCards.assignCard') : headerTitleAddCards;
+    const onOpenBankConnectionFlow = useCallback(() => {
+        if (!url) {
+            return;
+        }
+        customWindow = openBankConnection(url);
+    }, [url]);
+
+    useEffect(() => {
+        if (!policyID || !isBlockedToAddNewFeeds || feed) {
+            return;
+        }
+        Navigation.navigate(ROUTES.WORKSPACE_UPGRADE.getRoute(policyID, CONST.UPGRADE_FEATURE_INTRO_MAPPING.companyCards.alias, ROUTES.WORKSPACE_COMPANY_CARDS.getRoute(policyID)), {
+            forceReplace: true,
+        });
+    }, [isBlockedToAddNewFeeds, policyID, feed]);
+
+    const handleBackButtonPress = () => {
+        customWindow?.close();
+
+        // Handle assign card flow
+        if (feed) {
+            Navigation.goBack();
+            return;
+        }
+
+        // Handle add new card flow
+        if (backTo) {
+            Navigation.goBack(backTo);
+            return;
+        }
+        setAddNewCompanyCardStepAndData({step: CONST.COMPANY_CARDS.STEP.SELECT_BANK});
+    };
 
     const CustomSubtitle = (
         <Text style={[styles.textAlignCenter, styles.textSupporting]}>
-            {bankName && translate(`workspace.moreFeatures.companyCards.pendingBankDescription`, bankDisplayName ?? bankName)}
+            {bankName && translate(`workspace.moreFeatures.companyCards.pendingBankDescription`, addNewCard?.data?.plaidConnectedFeedName ?? bankName)}
             <TextLink onPress={onOpenBankConnectionFlow}>{translate('workspace.moreFeatures.companyCards.pendingBankLink')}</TextLink>.
         </Text>
     );
 
-    const getContent = () => {
-        if (isRefreshComplete) {
-            return (
-                <ConfirmationPage
-                    heading={translate('workspace.moreFeatures.companyCards.refreshConnectionSuccess')}
-                    description={translate('workspace.moreFeatures.companyCards.refreshConnectionSuccessDescription')}
-                    shouldShowButton
-                    buttonText={translate('common.buttonConfirm')}
-                    onButtonPress={() => Navigation.dismissModal()}
-                />
-            );
+    useEffect(() => {
+        if ((!url && !isPlaid) || isOffline || isNewFeedHasError || isAllFeedsResultLoading || (isBlockedToAddNewFeeds && !feed)) {
+            return;
         }
+
+        // Handle assign card flow
+        if (feed) {
+            if (!isFeedExpired) {
+                customWindow?.close();
+                if (isFeedConnectionBroken) {
+                    updateBrokenConnection();
+                    Navigation.closeRHPFlow();
+                    return;
+                }
+                setAssignCardStepAndData({
+                    currentStep: assignCard?.cardToAssign?.dateOption ? CONST.COMPANY_CARD.STEP.CONFIRMATION : CONST.COMPANY_CARD.STEP.ASSIGNEE,
+                    isEditing: false,
+                });
+                return;
+            }
+            if (isPlaid) {
+                return;
+            }
+            if (url) {
+                customWindow = openBankConnection(url);
+                return;
+            }
+        }
+
+        // Handle add new card flow
+        if (isNewFeedConnected) {
+            setShouldBlockWindowOpen(true);
+            customWindow?.close();
+            if (newFeed) {
+                updateSelectedFeed(newFeed, policyID);
+            }
+
+            // Direct feeds (except those added via Plaid) are created with default statement period end date.
+            // Redirect the user to set a custom date.
+            if (policyID && !isPlaid) {
+                setAddNewCompanyCardStepAndData({
+                    step: CONST.COMPANY_CARDS.STEP.SELECT_DIRECT_STATEMENT_CLOSE_DATE,
+                });
+            } else {
+                Navigation.closeRHPFlow();
+                Navigation.navigate(ROUTES.WORKSPACE_COMPANY_CARDS.getRoute(policyID), {forceReplace: true});
+            }
+            return;
+        }
+        if (!shouldBlockWindowOpen) {
+            if (isPlaid) {
+                onImportPlaidAccounts();
+                return;
+            }
+            if (url) {
+                customWindow = openBankConnection(url);
+            }
+        }
+    }, [
+        isNewFeedConnected,
+        isAllFeedsResultLoading,
+        shouldBlockWindowOpen,
+        isBlockedToAddNewFeeds,
+        newFeed,
+        policyID,
+        url,
+        feed,
+        isFeedExpired,
+        isOffline,
+        assignCard?.cardToAssign?.dateOption,
+        isPlaid,
+        onImportPlaidAccounts,
+        isFeedConnectionBroken,
+        updateBrokenConnection,
+        isNewFeedHasError,
+    ]);
+
+    const getContent = () => {
         if (isNewFeedHasError) {
             return (
                 <WorkspaceCompanyCardsErrorConfirmation
@@ -131,7 +232,7 @@ function BankConnection({policyID: policyIDFromProps, feed, route, isRefreshConn
 
     return (
         <ScreenWrapper
-            testID={isRefreshComplete ? 'RefreshCardFeedConnectionSuccess' : 'BankConnection'}
+            testID="BankConnection"
             enableEdgeToEdgeBottomSafeAreaPadding
         >
             <HeaderWithBackButton
