--- src/pages/workspace/companyCards/BankConnection/openBankConnection/index.native.ts
@@ -1,6 +1,0 @@
-const handleOpenBankConnectionFlow = () => {
-    // No-op for native
-    return null;
-};
-
-export default handleOpenBankConnectionFlow;
