+++ src/pages/workspace/companyCards/addNew/PlaidConnectionStep.tsx
@@ -27,17 +27,7 @@ import type {CompanyCardFeedWithDomainID} from '@src/types/onyx';
 import type {CardFeedWithNumber} from '@src/types/onyx/CardFeeds';
 import {isEmptyObject} from '@src/types/utils/EmptyObject';
 
-function PlaidConnectionStep({
-    feed,
-    policyID,
-    onExit,
-    isRefreshConnectionFlow,
-}: {
-    feed?: CompanyCardFeedWithDomainID;
-    policyID?: string;
-    onExit?: () => void;
-    isRefreshConnectionFlow?: boolean;
-}) {
+function PlaidConnectionStep({feed, policyID, onExit}: {feed?: CompanyCardFeedWithDomainID; policyID?: string; onExit?: () => void}) {
     const {translate} = useLocalize();
     const styles = useThemeStyles();
     const [addNewCard] = useOnyx(ONYXKEYS.ADD_NEW_COMPANY_CARD);
@@ -228,7 +218,7 @@ function PlaidConnectionStep({
             shouldEnableMaxHeight
         >
             <HeaderWithBackButton
-                title={translate(isRefreshConnectionFlow ? 'workspace.moreFeatures.companyCards.assignNewCards' : 'workspace.companyCards.addCards')}
+                title={translate('workspace.companyCards.addCards')}
                 onBackButtonPress={handleBackButtonPress}
             />
             {isPlaidDisabled ? (
