+++ src/pages/workspace/withPolicy.tsx
@@ -53,7 +53,6 @@ type PolicyRouteName =
     | typeof SCREENS.WORKSPACE.RULES
     | typeof SCREENS.WORKSPACE.EXPENSIFY_CARD_ISSUE_NEW
     | typeof SCREENS.WORKSPACE.COMPANY_CARDS_BROKEN_CARD_FEED_CONNECTION
-    | typeof SCREENS.WORKSPACE.COMPANY_CARDS_REFRESH_CARD_FEED_CONNECTION
     | typeof SCREENS.WORKSPACE.ACCOUNTING.CLAIM_OFFER
     | typeof SCREENS.WORKSPACE.TIME_TRACKING;
